LayoutTests/http/tests/websocket/tests/hybi/handshake-error.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Handshake error test");

window.jsTestIsAsync = true;
if (window.layoutTestController)
    layoutTestController.overridePreference("WebKitHixie76WebSocketProtocolEnabled", 0);

function endTest()
{
    clearTimeout(timeoutID);
    finishJSTest();
}

var ws = new WebSocket("ws://127.0.0.1:8880/websocket/tests/hybi/handshake-error");

ws.onopen = function()
{
    testFailed("Unexpectedly Connected.");
};

ws.onmessage = function(messageEvent)
{
    testFailed("Unexpectedly Received: '" + messageEvent.data + "'");
};

ws.onclose = function()
{
    debug("Closed.");
    shouldBe("ws.readyState", "3")
    endTest();
};

function timeOutCallback()
{
    testFailed("Timed out in state: " + ws.readyState);
    endTest();
}

var timeoutID = setTimeout(timeOutCallback, 3000);

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
