LayoutTests/http/tests/websocket/tests/hybi/handshake-error-expected.txt

CONSOLE MESSAGE: line 0: No response code found: ThisWillCauseHandshakeError
Handshake error test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Closed.
PASS ws.readyState is 3
PASS successfullyParsed is true

TEST COMPLETE
